refactor: separate application logic and presentation in posts page
